Refactor testing and coverage configuration
This commit updates the `.runsettings` file to specify a deterministic output directory for coverage results. It modifies the `global.json` to downgrade the `ktsu.Sdk` versions for better compatibility. The GitHub Actions workflow is enhanced by removing the installation step for `dotnet-coverage` and updating the SonarQube scanner commands for improved coverage reporting. Additionally, the `Invoke-DotNetTest` function is refactored to accept parameters for configuration and coverage output path, ensuring better flexibility and organization of test results. The `update-winget-manifests.ps1` script is also improved to better identify library and executable projects.